cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy yf225 chenyang78 kadeng muchulee8 ColinPeppler amjames desertfire chauhang aakhundov

[ghstack-poisoned]

+++ aten/src/ATen/native/cuda/DistanceKernel.cu
@@ -154,7 +154,7 @@ __global__ static void cdist_backward_kernel_cuda_impl(scalar_t * buffer, const 
 template <typename scalar_t, typename F>
 __global__ static void pdist_backward_kernel_cuda_impl(scalar_t * buffer, const scalar_t * grad, const scalar_t * self, const scalar_t * dist, int64_t gs, const int64_t n, const int64_t m, const int64_t combs, const scalar_t p,
                                                        const double n2, const double n2_squared_minus_1) {
-  const int64_t k = blockIdx.x * blockDim.x + threadIdx.x;
+  const int64_t k = ((int64_t) blockIdx.x) * blockDim.x + threadIdx.x;
   const int init = blockIdx.y * blockDim.y + threadIdx.y;
   const int stride = blockDim.y * gridDim.y;
